move logic to check db before server starts from bootloader into databaservice
